Avoid apps getting stuck waiting on a response from broken WS
When an error occurs (bad gateway or whatever), this function returns,
logs an error, but the error response is usually ignored. To address
this, and to keep in line with convention that the sender closes the
channel, this change makes sure that on exiting we close the channel, so
it's receivers will know that they shouldn't expect any other message
(thus can restart the listener, for example).